requirements.txt
fastapi
uvicorn
requests
python-telegram-bot
sib-api-v3-sdk
supabase